refactor: just moved the borrow page users in the components folder
